+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -350,8 +350,8 @@ public final class Realm extends BaseRealm {
         for (int i = 0; i < json.length(); i++) {
             try {
                 configuration.getSchemaMediator().createOrUpdateUsingJsonObject(clazz, this, json.getJSONObject(i), false);
-            } catch (Exception e) {
-                throw new RealmException("Could not map Json", e);
+            } catch (JSONException e) {
+                throw new RealmException("Could not map JSON", e);
             }
         }
     }
@@ -378,8 +378,8 @@ public final class Realm extends BaseRealm {
         for (int i = 0; i < json.length(); i++) {
             try {
                 configuration.getSchemaMediator().createOrUpdateUsingJsonObject(clazz, this, json.getJSONObject(i), true);
-            } catch (Exception e) {
-                throw new RealmException("Could not map Json", e);
+            } catch (JSONException e) {
+                throw new RealmException("Could not map JSON", e);
             }
         }
     }
@@ -402,7 +402,7 @@ public final class Realm extends BaseRealm {
         JSONArray arr;
         try {
             arr = new JSONArray(json);
-        } catch (Exception e) {
+        } catch (JSONException e) {
             throw new RealmException("Could not create JSON array from string", e);
         }
 
@@ -526,8 +526,8 @@ public final class Realm extends BaseRealm {
 
         try {
             return configuration.getSchemaMediator().createOrUpdateUsingJsonObject(clazz, this, json, false);
-        } catch (Exception e) {
-            throw new RealmException("Could not map Json", e);
+        } catch (JSONException e) {
+            throw new RealmException("Could not map JSON", e);
         }
     }
 
@@ -555,7 +555,7 @@ public final class Realm extends BaseRealm {
             handlerController.addToRealmObjects(realmObject);
             return realmObject;
         } catch (JSONException e) {
-            throw new RealmException("Could not map Json", e);
+            throw new RealmException("Could not map JSON", e);
         }
     }
 
@@ -578,7 +578,7 @@ public final class Realm extends BaseRealm {
         JSONObject obj;
         try {
             obj = new JSONObject(json);
-        } catch (Exception e) {
+        } catch (JSONException e) {
             throw new RealmException("Could not create Json object from string", e);
         }
 
@@ -609,7 +609,7 @@ public final class Realm extends BaseRealm {
         JSONObject obj;
         try {
             obj = new JSONObject(json);
-        } catch (Exception e) {
+        } catch (JSONException e) {
             throw new RealmException("Could not create Json object from string", e);
         }
 
@@ -1527,7 +1527,7 @@ public final class Realm extends BaseRealm {
         try {
             return realmFile.getCanonicalPath();
         } catch (IOException e) {
-            throw new RealmException("Could not resolve the canonical path to the Realm file: " + realmFile.getAbsolutePath());
+            throw new RealmIOException("Could not resolve the canonical path to the Realm file: " + realmFile.getAbsolutePath());
         }
     }