+++ realm/realm-annotations-processor/src/test/resources/io/realm/NullTypesRealmProxy.java
@@ -484,7 +484,7 @@ public class NullTypesRealmProxy extends NullTypes
             proxyState.getRow$realm().nullifyLink(columnInfo.fieldObjectNullIndex);
             return;
         }
-        if (!value.isValid()) {
+        if (!RealmObject.isValid(value)) {
             throw new IllegalArgumentException("'value' is not a valid managed object.");
         }
         if (((RealmObjectProxy)value).realmGet$proxyState().getRealm$realm() != proxyState.getRealm$realm()) {