+++ realm/realm-library/src/androidTest/java/io/realm/RealmAsyncQueryTests.java
@@ -43,7 +43,6 @@ import io.realm.entities.Owner;
 import io.realm.instrumentation.MockActivityManager;
 import io.realm.internal.async.RealmThreadPoolExecutor;
 import io.realm.internal.log.RealmLog;
-import io.realm.proxy.HandlerProxy;
 import io.realm.rule.RunInLooperThread;
 import io.realm.rule.RunTestInLooperThread;
 import io.realm.rule.TestRealmConfigurationFactory;