+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * All JSON methods on Realm now only wraps JSONException in RealmException. All other Exceptions are thrown as they are.
 * Removed `HandlerController` from the public API.
 * Marked all methods on `RealmObject` and all public classes final (#1594).
+* Removed constructor of `RealmAsyncTask` from the public API8 (#1594).
 
 ### Deprecated