+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ### Breaking changes
 
 * All JSON methods on Realm now only wraps JSONException in RealmException. All other Exceptions are thrown as they are.
-* Removed `HandlerController` from the public API.
 * Marked all methods on `RealmObject` and all public classes final (#1594).
+* Removed `BaseRealm` from the public API.
+* Removed `HandlerController` from the public API.
 * Removed constructor of `RealmAsyncTask` from the public API8 (#1594).
 
 ### Deprecated